Fix bug preventing the Bottom Navigation bar to detect clicks
This commit removes `setupContentBottomOffset()` from `onAttach()` of BaseFramesFragment.kt because view and getView() are both null at this point. Additionally, we're getting the measured height of `bottomNavigation` only if one of the views on which the padding should be applied isn't null. This prevents the initialization of `bottomNavigation` too early.